requirements.txt
django>=4.2
djangorestframework
drf-spectacular
psycopg2-binary
python-decouple
django-cors-headers
pip install django djangorestframework drf-spectacular django-cors-headers python-decouple